# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactResult {
  ok: boolean;
}

export interface NewsletterInput {
  email: string;
}

export interface NewsletterResult {
  ok: boolean;
}

export interface ErrorResult {
  error: string;
}