# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact and newsletter operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact message
      description: Saves a contact form submission (name, email, message)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /newsletter:
    post:
      operationId: subscribeNewsletter
      tags: [contact]
      summary: Subscribe to newsletter
      description: Adds an email to the newsletter subscribers list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "201":
          description: Subscribed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterResult"
        "409":
          description: Email already subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        message:
          type: string
          minLength: 1

    ContactResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    NewsletterInput:
      type: object
      required: [email]
      properties:
        email:
          type: string

    NewsletterResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    ErrorResult:
      type: object
      required: [error]
      properties:
        error:
          type: string